Finalized on newcoh, start new Pnc runs
After extensive tests using DCB, newcoh is finalized to be type-a,
rf=0.5 (discontinuous bilinear law)
Improved Pnc60clt-r1 model. A refined region is partitioned, such that
a uniform fine mesh is used in that region. Meshes of supporting parts
are consistent with the fnm part over the refined region. Loading line
is constrained from moving in the x and y direction.
Material properties are modified. Penalty stiffness is calculated based
on the formula in 2013 OHT, so it's much reduced. Toughness values from
Krueger are used. Delamination strengths are the same as those of
matrix cracks, thus not giving delamination an advantage.