bug fix to preproc and main
preproc: add treatment to 'internal' nset and elset in fnm part
main:
previously, the predelam is modeled by simply not integrating the
cohesive subelem in the predelam elset. This is NOT right, as in
compression, interpenetration would occur. The remedy is the following:
- add a pre crack procedure in the cohesive_material_module which would
reduce the strength and toughness of the material to SMALLNUM
- modify the fBrickLam module to use the pre crack procedure, such that
a pre cracked cohesive material is passed in the integration procedure
of the cohesive sub element in the predelam elset